requirements.txt
adafruit-circuitpython-ssd1306
pillow
psutil